+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -2,22 +2,31 @@ import { argentWallet } from './argentWallet/argentWallet';
 import { bitskiWallet } from './bitskiWallet/bitskiWallet';
 import { braveWallet } from './braveWallet/braveWallet';
 import { coinbaseWallet } from './coinbaseWallet/coinbaseWallet';
+import { dawnWallet } from './dawnWallet/dawnWallet';
 import { enkryptWallet } from './enkryptWallet/enkryptWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
 import { ledgerWallet } from './ledgerWallet/ledgerWallet';
 import { metaMaskWallet } from './metaMaskWallet/metaMaskWallet';
 import { mewWallet } from './mewWallet/mewWallet';
+import { okxWallet } from './okxWallet/okxWallet';
 import { omniWallet } from './omniWallet/omniWallet';
+import { phantomWallet } from './phantomWallet/phantomWallet';
+import { rabbyWallet } from './rabbyWallet/rabbyWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
+import { safeWallet } from './safeWallet/safeWallet';
+import { tahoWallet } from './tahoWallet/tahoWallet';
 import { trustWallet } from './trustWallet/trustWallet';
 import { walletConnectWallet } from './walletConnectWallet/walletConnectWallet';
+import { xdefiWallet } from './xdefiWallet/xdefiWallet';
+import { zerionWallet } from './zerionWallet/zerionWallet';
 
 export {
   argentWallet,
   bitskiWallet,
   braveWallet,
   coinbaseWallet,
+  dawnWallet,
   enkryptWallet,
   imTokenWallet,
   injectedWallet,
@@ -25,7 +34,14 @@ export {
   metaMaskWallet,
   mewWallet,
   omniWallet,
+  phantomWallet,
+  rabbyWallet,
   rainbowWallet,
+  safeWallet,
+  tahoWallet,
   trustWallet,
+  okxWallet,
   walletConnectWallet,
+  xdefiWallet,
+  zerionWallet,
 };