+++ packages/rainbowkit/src/wallets/walletConnectors/metaMaskWallet/metaMaskWallet.ts
@@ -7,37 +7,31 @@ import { Wallet } from '../../Wallet';
 import { getWalletConnectConnector } from '../../getWalletConnectConnector';
 
 export interface MetaMaskWalletOptions {
+  projectId?: string;
   chains: Chain[];
 }
 
-function isMetaMask(ethereum: NonNullable<typeof window['ethereum']>) {
+function isMetaMask(ethereum?: typeof window['ethereum']): boolean {
   // Logic borrowed from wagmi's MetaMaskConnector
-  // https://github.com/tmm/wagmi/blob/main/packages/core/src/connectors/metaMask.ts
-  const isMetaMask = Boolean(ethereum.isMetaMask);
-
-  if (!isMetaMask) {
-    return false;
-  }
-
+  // https://github.com/wagmi-dev/references/blob/main/packages/connectors/src/metaMask.ts
+  const isMetaMask = !!ethereum?.isMetaMask;
+  if (!isMetaMask) return false;
   // Brave tries to make itself look like MetaMask
   // Could also try RPC `web3_clientVersion` if following is unreliable
-  if (ethereum.isBraveWallet && !ethereum._events && !ethereum._state) {
+  if (ethereum.isBraveWallet && !ethereum._events && !ethereum._state)
     return false;
-  }
-
-  if (ethereum.isTokenPocket) {
-    return false;
-  }
-
-  if (ethereum.isTokenary) {
-    return false;
-  }
-
+  if (ethereum.isApexWallet) return false;
+  if (ethereum.isAvalanche) return false;
+  if (ethereum.isKuCoinWallet) return false;
+  if (ethereum.isPortal) return false;
+  if (ethereum.isTokenPocket) return false;
+  if (ethereum.isTokenary) return false;
   return true;
 }
 
 export const metaMaskWallet = ({
   chains,
+  projectId,
   ...options
 }: MetaMaskWalletOptions & MetaMaskConnectorOptions): Wallet => {
   const isMetaMaskInjected =
@@ -63,7 +57,7 @@ export const metaMaskWallet = ({
     },
     createConnector: () => {
       const connector = shouldUseWalletConnect
-        ? getWalletConnectConnector({ chains })
+        ? getWalletConnectConnector({ projectId, chains })
         : new MetaMaskConnector({
             chains,
             options,